Explicitly request "response_mode=fragment" in authorization requests
This parameter is specified by https://openid.net/specs/oauth-v2-multiple-response-types-1_0.html

Since this library deviates from the OAuth specification by unconditionally
expecting a fragment response, this expectation needs to be communicated to
the authorization server, to give it a chance to respond appropriately.
Otherwise, servers must hardcode the use of a fragment response, which may
prevent interoperability with other OAuth clients.